# src/app/api/services/offers/create/route.ts
import { NextRequest, NextResponse } from 'next/server';
import { CreateServiceOfferRequestSchema, ServiceOfferResponseSchema } from '@/lib/schemas/service';
import { validateRequest } from '@/lib/utils/validation';
import { createSuccessResponse, createErrorResponse, COMMON_ERROR_CODES } from '@/lib/utils/response';
import { requireAuth, requireProfessionalOrAdmin } from '@/lib/auth/middleware';
import { createServiceOffer, getServiceOfferById } from '@/lib/db/queries/service';

/**
 * Create service offer
 * @description Create a new service offer for a posting
 * @body CreateServiceOfferRequestSchema
 * @response 201:ServiceOfferResponseSchema:Service offer created successfully
 * @responseSet auth
 * @security BearerAuth
 * @openapi
 */
export async function POST(request: NextRequest) {
  try {
    // Authenticate user (must be professional)
    const auth = requireProfessionalOrAdmin(request);
    if (!auth.success) {
      return auth.response;
    }

    const { userId } = auth.user;

    // Validate request body
    const validation = await validateRequest(request, CreateServiceOfferRequestSchema);
    if (!validation.success) {
      return NextResponse.json(validation.error, { status: 400 });
    }

    const {
      postingId,
      proposedPrice,
      description,
      estimatedDuration,
    } = validation.data;

    // Create service offer
    const offer = await createServiceOffer({
      postingId,
      professionalId: userId,
      proposedPrice,
      description,
      estimatedDuration,
    });

    // Get the created offer with relations
    const createdOffer = await getServiceOfferById(offer.id);

    // Prepare response data
    const responseData = {
      offer: {
        id: createdOffer.id,
        postingId: createdOffer.postingId,
        professionalId: createdOffer.professionalId,
        proposedPrice: createdOffer.proposedPrice,
        description: createdOffer.description,
        estimatedDuration: createdOffer.estimatedDuration,
        status: createdOffer.status,
        createdAt: createdOffer.createdAt.toISOString(),
        updatedAt: createdOffer.updatedAt.toISOString(),
      },
    };

    // Validate response
    const responseValidation = ServiceOfferResponseSchema.safeParse(responseData);
    if (!responseValidation.success) {
      return NextResponse.json(
        createErrorResponse(
          COMMON_ERROR_CODES.INTERNAL_ERROR,
          'Response validation failed'
        ),
        { status: 500 }
      );
    }

    return NextResponse.json(
      createSuccessResponse(responseValidation.data, 'Service offer created successfully'),
      { status: 201 }
    );

  } catch (error) {
    console.error('Create service offer error:', error);
    return NextResponse.json(
      createErrorResponse(
        COMMON_ERROR_CODES.INTERNAL_ERROR,
        'Failed to create service offer'
      ),
      { status: 500 }
    );
  }
}